pyproject.toml
[project]
name = "notion-week-report"
version = "0.1.0"
description = "自动周报生成工具 - 从 Notion 任务跟踪器自动生成周报并使用 DeepSeek 润色"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notion-client>=2.2.1",
    "openai>=1.55.0",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
    "httpx>=0.27.2",
    "pydantic>=2.10.2",
    "pyyaml>=6.0.2",
    "socksio>=1.0.0",
]

[project.scripts]
week-report = "notion_week_report.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_week_report"]
